pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "postura"
version = "0.1.0"
description = "Attack Surface Posture Agent — agentic security graph for your codebase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "neo4j>=5.17.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "tree-sitter>=0.22.0",
    "tree-sitter-python>=0.22.0",
    "httpx>=0.27.0",
    "anthropic>=0.25.0",
    "langgraph>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "gitpython>=3.1.0",
    "pip-audit>=2.7.0",
    "semgrep>=1.70.0",
    "bandit>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "testcontainers[neo4j]>=4.0.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
